% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexternalinstrument.R
\name{externalinstrument.data.frame}
\alias{externalinstrument.data.frame}
\title{Identify the impulse response for a VAR (using the residuals from the VAR), using a high frequency instrument.}
\usage{
\method{externalinstrument}{data.frame}(res, instrument, dependent, p)
}
\arguments{
\item{res}{A data frame containing the reduced form residuals from your vAR.}

\item{instrument}{A list containing the data for the instrument. Should be same length as the data frame of residuals.}

\item{dependent}{Which variable in your var are you instrumenting (as a string).}

\item{p}{(Integer) How many lags does your var have.}
}
\description{
Identify the impulse response for a VAR (using the residuals from the VAR), using a high frequency instrument.
}
