% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexternalinstrument.R
\name{externalinstrument}
\alias{externalinstrument}
\title{Identify the impulse response for a VAR (using the VAR estimated from the vars package), using a high frequency instrument.}
\usage{
externalinstrument(var, instrument, dependent, p)
}
\arguments{
\item{var}{A varest var, or a dataframe of reduced form residuals.}

\item{instrument}{A list containing the data for the instrument. Should be same length as the estimation sample.}

\item{dependent}{Which variable in your var are you instrumenting (as a string).}

\item{p}{(Integer) How many lags does your var have (only needed if supplying a dataframe instead of a varest).}
}
\description{
Identify the impulse response for a VAR (using the VAR estimated from the vars package), using a high frequency instrument.
}
\examples{
library(vars)
library(varexternalinstrument)
data(GKdata)
gkvar <- VAR(GKdata[, c("logip", "logcpi", "gs1", "ebp")], p = 12, type = "const")
shockcol <- externalinstrument(gkvar, GKdata$ff4_tc, "gs1")

}
