% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexternalinstrument.R
\name{externalinstrument.varest}
\alias{externalinstrument.varest}
\title{Identify the impulse response for a VAR (using the VAR estimated from the vars package), using a high frequency instrument.}
\usage{
\method{externalinstrument}{varest}(var, instrument, dependent)
}
\arguments{
\item{var}{A varest object (resulting from a var estimation from the vars package).}

\item{instrument}{A list containing the data for the instrument. Should be same length as the estimation sample.}

\item{dependent}{Which variable in your var are you instrumenting (as a string).}
}
\description{
Identify the impulse response for a VAR (using the VAR estimated from the vars package), using a high frequency instrument.
}
\examples{
library(vars)
library(varexternalinstrument)
data(GKdata)
gkvar <- VAR(GKdata[, c("logip", "logcpi", "gs1", "ebp")], p = 12, type = "const")
shockcol <- externalinstrument(gkvar, GKdata$ff4_tc, "gs1")

}
