% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varexternalinstrument.R
\docType{data}
\name{GKdata}
\alias{GKdata}
\title{Replication data from Gertler and Karadi (2015).}
\description{
Replication data from Gertler and Karadi (2015).
}
\references{
\url{https://www.aeaweb.org/articles?id=10.1257/mac.20130329}
}
\keyword{data}
